move plan runner into new catkin package
