--- src/catkin_projects/drake_iiwa_sim/src/task_space_trajectory_plan.cc
@@ -1,2 +1,0 @@
-#include <robot_plan_runner/task_space_trajectory_plan.h>
-
