+++ src/catkin_projects/drake_robot_control/.idea/drake_robot_control.iml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module classpath="CMake" type="CPP_MODULE" version="4" />
