+++ src/catkin_projects/drake_robot_control/src/joint_space_trajectory_plan.cc
@@ -1,0 +1,1 @@
+#include <drake_robot_control/joint_space_trajectory_plan.h>
