+++ src/catkin_projects/drake_robot_control/src/plan_runner.cc
@@ -1,4 +1,4 @@
-#include <robot_plan_runner/plan_runner.h>
+#include <drake_robot_control/plan_runner.h>
 #include <yaml-cpp/yaml.h>
 
 namespace drake {
