+++ src/catkin_projects/drake_robot_control/src/task_space_trajectory_plan.cc
@@ -1,0 +1,1 @@
+#include <drake_robot_control/task_space_trajectory_plan.h>
